python/requirements.txt
faiss-cpu>=1.13.0
googletrans>=4.0.2
langdetect>=1.0.9
numba>=0.62.1
openai>=2.8.1
pymupdf>=1.26.6
pypdf2>=3.0.1
pyyaml>=6.0
sentence-splitter>=1.4
sentence-transformers>=5.1.2